Move logic for applying deps into executors
This way, each executor can decide how to treat each kind of dep. For
instance, a "dask" or "local" executor could ignore ModuleDeps.

Also change Electron signature to accept a `deps` dictionary, e.g.

@ct.electron(deps={"bash": BashDeps(...), "pip": PipDeps(...)})
