+++ covalent/_shared_files/defaults.py
@@ -109,7 +109,7 @@ def set_executor() -> dict:
 # Going forward we may only want to return the executor field of DEFAULT_CONSTRAINT_VALUES
 # The rest of those parameters will now be in this dictionary
 _DEFAULT_CONSTRAINT_VALUES = set_executor()
-_DEFAULT_CONSTRAINT_VALUES["bash_deps"] = None
+_DEFAULT_CONSTRAINT_VALUES["deps"] = {}
 
 _DEFAULT_CONSTRAINTS_DEPRECATED = {
     "schedule": False,
