+++ covalent/executor/base.py
@@ -40,26 +40,6 @@ app_log = logger.app_log
 log_stack_info = logger.log_stack_info
 
 
-def wrapper_fn(function: TransportableObject, pre_cmds: [], *args, **kwargs):
-    """Wrapper for serialized callable.
-
-    Execute preparatory shell commands before deserializing and
-    running the callable. This is the actual function to be sent to
-    the various executors.
-
-    """
-    for cmd in pre_cmds:
-        proc = subprocess.run(cmd, stdin=subprocess.DEVNULL, capture_output=True)
-        try:
-            proc.check_returncode()
-        except subprocess.CalledProcessError as ex:
-            raise ex
-
-    fn = function.get_deserialized()
-    output = fn(*args, **kwargs)
-    return output
-
-
 class BaseExecutor(ABC):
     """
     Base executor class to be used for defining any executor
@@ -147,6 +127,7 @@ class BaseExecutor(ABC):
         function: TransportableObject,
         args: List,
         kwargs: Dict,
+        deps: Dict,
         dispatch_id: str,
         results_dir: str,
         node_id: int = -1,
