+++ covalent/executor/executor_plugins/dask.py
@@ -27,6 +27,7 @@ This is a plugin executor module; it is loaded if found and properly structured.
 
 import io
 import os
+import subprocess
 from contextlib import redirect_stderr, redirect_stdout
 from typing import Any, Dict, List
 
@@ -38,7 +39,7 @@ from covalent._shared_files import logger
 from covalent._shared_files.config import get_config
 from covalent._shared_files.util_classes import DispatchInfo
 from covalent._workflow.transport import TransportableObject
-from covalent.executor import BaseExecutor, wrapper_fn
+from covalent.executor import BaseExecutor
 
 # The plugin class name must be given by the executor_plugin_name attribute:
 executor_plugin_name = "DaskExecutor"
@@ -53,6 +54,24 @@ _EXECUTOR_PLUGIN_DEFAULTS = {
         os.environ.get("XDG_CACHE_HOME") or os.path.join(os.environ["HOME"], ".cache"), "covalent"
     ),
 }
+
+
+def wrapper_fn(function: TransportableObject, pre_cmds: [], *args, **kwargs):
+    """Wrapper for serialized callable.
+
+    Execute preparatory shell commands before deserializing and
+    running the callable. This is the actual function to be sent to
+    the various executors.
+
+    """
+    for cmd in pre_cmds:
+        proc = subprocess.run(
+            cmd, stdin=subprocess.DEVNULL, shell=True, capture_output=True, check=True, text=True
+        )
+
+    fn = function.get_deserialized()
+    output = fn(*args, **kwargs)
+    return output
 
 
 class DaskExecutor(BaseExecutor):
@@ -87,6 +106,7 @@ class DaskExecutor(BaseExecutor):
         function: TransportableObject,
         args: List,
         kwargs: Dict,
+        deps: Dict,
         dispatch_id: str,
         results_dir: str,
         node_id: int = -1,
@@ -114,7 +134,12 @@ class DaskExecutor(BaseExecutor):
 
         fn_version = function.python_version
 
-        new_args = [function, []]
+        pre_cmds = []
+        if "bash" in deps:
+            bash_deps = deps["bash"]
+            pre_cmds.extend(bash_deps.apply())
+
+        new_args = [function, pre_cmds]
         for arg in args:
             new_args.append(arg)
