+++ covalent_dispatcher/_core/execution.py
@@ -44,6 +44,7 @@ from covalent._shared_files.defaults import (
     sublattice_prefix,
     subscript_prefix,
 )
+from covalent._workflow.deps import Deps
 from covalent._workflow.lattice import Lattice
 from covalent.executor import _executor_manager
 from covalent_ui import result_webhook
@@ -175,6 +176,7 @@ def _run_task(
     inputs: Dict,
     serialized_callable: Any,
     selected_executor: Any,
+    deps: Dict[str, Deps],
     node_name: str,
 ) -> None:
     """
@@ -223,6 +225,7 @@ def _run_task(
                 function=serialized_callable,
                 args=inputs["args"],
                 kwargs=inputs["kwargs"],
+                deps=deps,
                 dispatch_id=dispatch_id,
                 results_dir=results_dir,
                 node_id=node_id,
@@ -328,6 +331,10 @@ def _run_planned_workflow(result_object: Result, thread_pool: ThreadPoolExecutor
                 node_id, "metadata"
             )["executor"]
 
+            deps = result_object.lattice.transport_graph.get_node_value(node_id, "metadata")[
+                "deps"
+            ]
+
             update_node_result(
                 generate_node_result(
                     node_id=node_id,
@@ -345,6 +352,7 @@ def _run_planned_workflow(result_object: Result, thread_pool: ThreadPoolExecutor
                 serialized_callable=serialized_callable,
                 selected_executor=pickle.dumps(selected_executor),
                 node_name=node_name,
+                deps=deps,
                 inputs=pickle.dumps(task_input),
             )
